Initial commit of version and branch name change. Update Readme. This branch is to make all Point coordinates double precision through the rendering pipeline, for robustness, especially when re-importing past render results.

+++ Objtree.h
@@ -121,9 +121,9 @@ typedef struct Object
 typedef struct Point
 {
     struct Object   hdr;            // Header
-    float           x;              // Coordinates
-    float           y;
-    float           z;
+    double          x;              // Coordinates
+    double          y;
+    double          z;
     BOOL            moved;          // When a point is moved, this is set TRUE. This stops shared
                                     // points from being moved twice.
     float           decay;          // Decay factor for smooth moves.
